+++ src/components/chat/ChatInput.vue
@@ -1,5 +1,5 @@
 <template>
-  <div>
+  <div class="vdb-c-border-t vdb-c-border-kilvish-400 vdb-c-p-16">
     <div
       v-if="showAgentList"
       class="vdb-c-absolute vdb-c-z-50 vdb-c-w-full vdb-c--translate-y-full vdb-c-transform vdb-c-px-18"
@@ -22,97 +22,112 @@
       </div>
     </div>
     <div
-      class="vdb-c-absolute vdb-c-left-0 vdb-c-top-0 vdb-c-h-full vdb-c-w-full vdb-c-px-10 vdb-c-py-8 md:vdb-c-p-18"
+      :class="[
+        'vdb-c-border vdb-c-p-8',
+        inputFocused ? 'vdb-c-border-kilvish-600' : 'vdb-c-border-kilvish-400',
+        isExpanded ? 'vdb-c-rounded-20' : 'vdb-c-rounded-[50px]',
+      ]"
     >
-      <div class="vdb-c-br-50 vdb-c-relative vdb-c-h-full vdb-c-w-full">
-        <form
-          class="vdb-c-br-50 vdb-c-flex vdb-c-h-full vdb-c-w-full vdb-c-items-center vdb-c-justify-between vdb-c-overflow-hidden vdb-c-border vdb-c-border-solid vdb-c-bg-white"
-          :class="{
-            'vdb-c-border-kilvish-600': inputFocused,
-            'vdb-c-border-kilvish-400': !inputFocused,
-          }"
-          autocomplete="off"
-          @submit="handleSubmit"
+      <div
+        class="vdb-c-m-8 vdb-c-mb-14 vdb-c-flex vdb-c-flex-col vdb-c-gap-8"
+        v-if="imageAttachments.length > 0"
+      >
+        <ChatInputImagePreview
+          v-for="attachment in imageAttachments"
+          :key="attachment.key"
+          :attachment="attachment"
+          @remove="removeAttachment(attachment)"
+        />
+      </div>
+
+      <div class="vdb-c-flex vdb-c-items-center vdb-c-gap-8">
+        <!-- Upload Button -->
+        <div
+          :class="[
+            'vdb-c-ml-4 vdb-c-cursor-pointer vdb-c-px-8',
+            isExpanded ? 'vdb-c-self-end vdb-c-py-10' : 'vdb-c-py-2',
+          ]"
         >
-          <div>
-            <div
-              :class="[
-                'vdb-c-ml-6 vdb-c-hidden vdb-c-items-center vdb-c-gap-24 vdb-c-rounded-[42px] vdb-c-border vdb-c-border-[#BAE7BC] vdb-c-bg-[#E6F6E7] vdb-c-px-12 vdb-c-py-12 vdb-c-pr-16 vdb-c-text-black md:vdb-c-flex',
-                { 'vdb-c-animate-pulse': !contextData?.name },
-              ]"
-            >
-              <span
-                class="context-icon vdb-c-ml-12 vdb-c-h-8 vdb-c-w-8 vdb-c-rounded-full vdb-c-border vdb-c-border-white"
-              >
-              </span>
-              <span
-                class="vdb-c-max-w-[160px] vdb-c-truncate vdb-c-text-sm vdb-c-font-bold"
-              >
-                {{ contextData?.name || "Loading..." }}
-              </span>
+          <label class="vdb-c-cursor-pointer">
+            <input
+              id="upload"
+              type="file"
+              accept="image/*"
+              class="vdb-c-hidden"
+              @change="handleFileUpload"
+            />
+            <PaperClipIcon />
+          </label>
+        </div>
+
+        <!-- Textarea -->
+        <textarea
+          ref="inputRef"
+          type="text"
+          class="vdb-c-chat-input vdb-c-max-h-[25vh] vdb-c-font-medium vdb-c-text-[#1D2736] vdb-c-placeholder-kilvish-500 vdb-c-outline-none focus:vdb-c-outline-none"
+          rows="1"
+          :placeholder="placeholder"
+          autocomplete="off"
+          :value="chatInput"
+          @input="handleInput"
+          @focus="inputFocused = true"
+          @blur="handleBlur"
+          @keydown="handleKeyDown"
+          @paste="handlePaste"
+        ></textarea>
+
+        <!-- Send Button -->
+        <div
+          class="vdb-c-flex vdb-c-items-center"
+          :class="{
+            'vdb-c-self-end': isExpanded,
+          }"
+        >
+          <button
+            :disabled="isInputDisabled"
+            class="vdb-c-font-sans vdb-c-hidden vdb-c-h-40 vdb-c-items-center vdb-c-justify-center vdb-c-rounded-full vdb-c-px-12 vdb-c-py-8 vdb-c-text-sm vdb-c-font-bold vdb-c-uppercase vdb-c-text-white vdb-c-transition md:vdb-c-flex"
+            :class="{
+              'vdb-c-cursor-not-allowed vdb-c-bg-kilvish-400 hover:vdb-c-bg-kilvish-400':
+                isInputDisabled,
+              'hover:vdb-c-orange-400 vdb-c-cursor-pointer vdb-c-bg-orange':
+                !isInputDisabled,
+            }"
+            @click="handleSubmit"
+            type="submit"
+          >
+            <EllipsesLoading v-if="chatLoading" />
+            <div v-else class="vdb-c-flex vdb-c-items-center vdb-c-pl-4">
+              <span class="vdb-c-inline">Send</span>
+              <chat-enter-icon class-name="vdb-c-ml-4" />
             </div>
-          </div>
-          <div class="vdb-c-relative vdb-c-h-full vdb-c-flex-grow">
-            <textarea
-              ref="inputRef"
-              type="text"
-              class="vdb-c-chat-input vdb-c-h-full vdb-c-w-full vdb-c-bg-white vdb-c-pl-16 vdb-c-pr-8 vdb-c-pt-8 vdb-c-font-medium vdb-c-text-[#1D2736] vdb-c-placeholder-kilvish-500 focus:vdb-c-outline-none md:vdb-c-pt-[13px]"
-              name="prompt"
-              :placeholder="placeholder"
-              autocomplete="off"
-              :value="chatInput"
-              @input="handleInput"
-              @focus="inputFocused = true"
-              @blur="handleBlur"
-              @keydown="handleKeyDown"
-              style="
-                resize: none;
-                min-height: 40px;
-                max-height: 25vh;
-                box-sizing: border-box;
-              "
-            ></textarea>
-          </div>
-          <div class="vdb-c-flex vdb-c-items-center vdb-c-justify-end">
-            <button
-              :disabled="isInputDisabled"
-              class="vdb-c-font-sans vdb-c-mx-8 vdb-c-hidden vdb-c-h-40 vdb-c-cursor-pointer vdb-c-items-center vdb-c-justify-center vdb-c-rounded-full vdb-c-bg-orange vdb-c-px-12 vdb-c-py-8 vdb-c-text-sm vdb-c-font-bold vdb-c-uppercase vdb-c-text-white vdb-c-transition hover:vdb-c-bg-orange-400 md:vdb-c-flex"
-              :class="{
-                'vdb-c-cursor-not-allowed vdb-c-bg-kilvish-400 hover:vdb-c-bg-kilvish-400':
-                  charCount < 1,
-              }"
-              type="submit"
-            >
-              <EllipsesLoading v-if="chatLoading" />
-              <div v-else class="vdb-c-flex vdb-c-items-center vdb-c-pl-4">
-                <span class="vdb-c-inline">Send</span>
-                <chat-enter-icon class-name="vdb-c-ml-4" />
-              </div>
-            </button>
-            <button
-              class="vdb-c-mobile-send vdb-c-flex vdb-c-border-none vdb-c-bg-transparent vdb-c-p-8 vdb-c-pr-12 md:vdb-c-hidden"
-              :class="{
-                'vdb-c-text-kilvish-400': isInputDisabled,
-                'vdb-c-mobile-send vdb-c-text-others-nightwing':
-                  !isInputDisabled,
-              }"
-              type="submit"
-            >
-              <send-icon class-name="vdb-c-w-24 vdb-c-h-24" />
-            </button>
-          </div>
-        </form>
+          </button>
+          <button
+            class="vdb-c-mobile-send vdb-c-flex vdb-c-border-none vdb-c-bg-transparent vdb-c-p-8 vdb-c-pr-12 md:vdb-c-hidden"
+            :class="{
+              'vdb-c-bg-kilvish-400 vdb-c-text-kilvish-400 hover:vdb-c-bg-kilvish-400':
+                isInputDisabled,
+              'vdb-c-mobile-send vdb-c-text-others-nightwing': !isInputDisabled,
+            }"
+            @click="handleSubmit"
+            type="submit"
+          >
+            <send-icon class-name="vdb-c-w-24 vdb-c-h-24" />
+          </button>
+        </div>
       </div>
     </div>
   </div>
 </template>
 
 <script setup>
-import { ref, computed, watch } from "vue";
+import { ref, computed, watch, nextTick } from "vue";
 import { useVideoDBChat } from "../../context";
 import ChatEnterIcon from "../icons/ChatEnter.vue";
 import SendIcon from "../icons/Send.vue";
+import PaperClipIcon from "../icons/PaperClip.vue";
 import EllipsesLoading from "./elements/EllipsesLoading.vue";
+import ChatInputImagePreview from "./elements/ChatInputImagePreview.vue";
+import { v4 as uuidv4 } from "uuid";
 
 const props = defineProps({
   placeholder: {
@@ -129,7 +144,7 @@ const props = defineProps({
   },
 });
 
-const { chatInput, chatLoading } = useVideoDBChat();
+const { chatInput, chatAttachments, chatLoading } = useVideoDBChat();
 
 const emit = defineEmits(["on-submit", "on-change", "tag-agent"]);
 
@@ -140,6 +155,7 @@ const showAgentList = ref(false);
 const agentStartIndex = ref(-1);
 const agentQuery = ref("");
 const selectedAgentIndex = ref(0);
+const isTextBoxExpanded = ref(false);
 
 const focus = () => {
   inputRef.value.focus();
@@ -156,9 +172,22 @@ watch(filteredAgents, () => {
   selectedAgentIndex.value = 0;
 });
 
+const imageAttachments = computed(() =>
+  chatAttachments.filter((attachment) => attachment.type === "image"),
+);
+
 const isInputDisabled = computed(() => {
-  return chatLoading.value || charCount.value < 1;
+  const hasUploadingImages = imageAttachments.value.some(
+    (attachment) =>
+      attachment.upload_status === "uploading" ||
+      attachment.upload_status === "in_queue",
+  );
+  return chatLoading.value || charCount.value < 1 || hasUploadingImages;
 });
+
+const isExpanded = computed(
+  () => isTextBoxExpanded.value || chatAttachments.length > 0,
+);
 
 const handleInput = (e) => {
   const newValue = e.target.value;
@@ -174,6 +203,7 @@ const handleInput = (e) => {
   } else {
     showAgentList.value = false;
   }
+  adjustHeight(); // Adjust height dynamically
 };
 
 const resetTag = () => {
@@ -218,6 +248,7 @@ const handleKeyDown = (e) => {
       e.preventDefault();
     }
   } else if (e.shiftKey && e.key === "Enter") {
+    adjustHeight(); // Adjust height on Shift+Enter
     return;
   }
 };
@@ -241,14 +272,87 @@ watch(
   { immediate: true },
 );
 
-const handleSubmit = (e) => {
+const handlePaste = (event) => {
+  const clipboardData = event.clipboardData || window.clipboardData;
+  const items = Array.from(clipboardData.items);
+
+  const imageItem = items.find(
+    (item) => item.kind === "file" && item.type.startsWith("image/"),
+  );
+
+  if (imageItem) {
+    const file = imageItem.getAsFile();
+    if (file) {
+      const newImageAttachment = {
+        type: "image",
+        image_data: file,
+        key: uuidv4(),
+        upload: true,
+        upload_status: "in_queue",
+      };
+      attachAttachment(newImageAttachment);
+    }
+    event.preventDefault(); // Prevent default paste behavior
+  }
+};
+
+const handleFileUpload = (event) => {
+  const file = event.target.files[0];
+  if (file) {
+    const newImageAttachment = {
+      type: "image",
+      image_data: file,
+      key: uuidv4(),
+      upload: true,
+      upload_status: "in_queue",
+    };
+    attachAttachment(newImageAttachment);
+  }
+};
+
+const handleSubmit = async (e) => {
+  if (isInputDisabled.value) return;
   e.preventDefault();
   if (!showAgentList.value && chatInput.value.trim() !== "") {
-    emit("on-submit", chatInput.value);
+    emit("on-submit", {
+      text: chatInput.value,
+      images: imageAttachments.value,
+    });
     chatInput.value = "";
+    clearAllAttachments();
     charCount.value = 0;
     resetTag();
+    await nextTick();
+    adjustHeight();
   }
+};
+
+function adjustHeight() {
+  const textarea = inputRef.value;
+  textarea.style.height = "auto"; // Reset height to recalculate
+  textarea.style.height = textarea.scrollHeight + "px"; // Set height based on content
+  if (textarea.scrollHeight > 30) {
+    isTextBoxExpanded.value = true;
+  } else {
+    isTextBoxExpanded.value = false;
+  }
+}
+
+const attachAttachment = (attachment) => {
+  chatAttachments[0] = attachment;
+};
+
+const removeAttachment = (attachment) => {
+  const index = chatAttachments.findIndex(
+    (item) => item.key === attachment.key,
+  );
+  if (index !== -1) {
+    chatAttachments.splice(index, 1);
+  }
+};
+
+const clearAllAttachments = () => {
+  chatAttachments.splice(0, chatAttachments.length);
 };
 
 defineExpose({
@@ -257,26 +361,18 @@ defineExpose({
 </script>
 
 <style>
+/* Shared Styling */
 .vdb-c-br-50 {
   border-radius: 50px;
 }
 
+/* Chat Input */
 .vdb-c-chat-input {
   flex-grow: 1;
+  resize: none;
 }
 
 .vdb-c-chat-input::placeholder {
   font-weight: 400;
-}
-
-.vdb-c-chat-enter-button {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: 2fr 3fr;
-}
-
-.context-icon {
-  background: radial-gradient(circle, #5bc25f 0%, #0aa910 100%);
-  box-shadow: 0 0 6px 4px #0aa910;
 }
 </style>