+++ src/components/chat/ChatMessageContainer.vue
@@ -1,7 +1,7 @@
 <template>
   <div
-    class="vdb-c-relative vdb-c-h-auto vdb-c-w-full vdb-c-px-30"
-    :class="{ 'vdb-c-pt-16 vdb-c-min-h-full': isLastConv }"
+    class="vdb-c-relative vdb-c-h-auto vdb-c-w-full"
+    :class="{ 'vdb-c-min-h-full vdb-c-pt-16': isLastConv }"
   >
     <div
       v-for="(message, i) in messages"