+++ src/components/hooks/useChatInterface.js
@@ -3,6 +3,7 @@ import { ref, reactive } from "vue";
 export function useChatInterface() {
   const messageHandlers = {};
   const chatInput = ref("");
+  const chatAttachments = reactive([])
 
   const registerMessageHandler = (contentType, handler) => {
     messageHandlers[contentType] = handler;
@@ -14,6 +15,7 @@ export function useChatInterface() {
 
   return {
     chatInput,
+    chatAttachments,
     setChatInput,
     messageHandlers,
     registerMessageHandler,